+++ kubejs_scripts/1.20/dye_unifier_1.20-nether-expansion.js
@@ -1,0 +1,7 @@
+// Adds recipes that use items from The Outer End
+ServerEvents.recipes(e => {
+
+    // Cyan
+    e.recipes.thermal.centrifuge(["2x blue_dye", "2x light_blue_dye"], "netherexp:warphopper_fur", 2).energy(1800).id("dyeunifier:centrifuge_warphopper_fur")
+
+})